+++ packages/ocean-react/src/_shared/components/ListContainer/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ListContainer';
+export type { ListContainerProps } from './ListContainer';
+export type { ListContainerHighlight } from './ListContainer';